rename "part" to "part pattern"
