+++ packages/route-pattern/src/lib2/part-pattern/ast.ts
