+++ packages/route-pattern/src/lib2/part-pattern/variants.test.ts
