+++ packages/route-pattern/src/lib2/route-pattern/ast.ts
@@ -1,10 +1,10 @@
-import type * as Part from '../part/index.ts'
+import type * as PartPattern from '../part-pattern/index.ts'
 import type * as Search from './search.ts'
 
 export type AST = {
-  protocol: Part.AST | undefined
-  hostname: Part.AST | undefined
+  protocol: PartPattern.AST | undefined
+  hostname: PartPattern.AST | undefined
   port: string | undefined
-  pathname: Part.AST | undefined
+  pathname: PartPattern.AST | undefined
   search: Search.Constraints
 }
