+++ packages/route-pattern/src/lib2/route-pattern/variants.ts
@@ -1,5 +1,5 @@
 import type { AST } from './ast.ts'
-import * as Part from '../part/index.ts'
+import * as PartPattern from '../part-pattern/index.ts'
 
 type Tuple4<T> = [protocol: T, hostname: T, port: T, pathname: T]
 
@@ -9,9 +9,9 @@ type Variant = {
 }
 
 export function* variants(pattern: AST): Generator<Variant> {
-  let protocols = pattern.protocol ? Part.variants(pattern.protocol) : [undefined]
-  let hostnames = pattern.hostname ? Part.variants(pattern.hostname) : [undefined]
-  let pathnames = pattern.pathname ? Part.variants(pattern.pathname) : [undefined]
+  let protocols = pattern.protocol ? PartPattern.variants(pattern.protocol) : [undefined]
+  let hostnames = pattern.hostname ? PartPattern.variants(pattern.hostname) : [undefined]
+  let pathnames = pattern.pathname ? PartPattern.variants(pattern.pathname) : [undefined]
 
   for (let protocol of protocols) {
     for (let hostname of hostnames) {
